add new sockets pool code
With this changes we now rely on an atomic counter to know if we need to check
to the pool or not. It removes us one extra call to the gen_server if no
connections are available. Also we don't queue anymore the clients request: if
not socket is available in the pool, a new connection is made. It solves us
for some common supervision issues and makes the code more simple. Such
concurrency should be handle now in the application code.

While we are here we also introduces the following modules:
- hackney_sockets_sup, a supervisor to supervise pools
- hackney_sockets_server, to look at pool info and maintain the sockets
  registry. Other usages like maintaining a list of server sockets will also
be possible in near future.

Note: for now the new pool is not connected to hackney.