add hackney_socket:sendfile/2, hackney_socket:sendfile/5 functions
function similar to file:sendfile but for hackney_socket. Also
hackney_socket:sendfile/5 takes a `send_fun` options to set the function to
effectively send the data. Useful when you want to transform the data on the
fly or want to wrap it. The function like hackney_socket:send/2 takes 2
parameter: HackneySocket and Data. Default is to use hackney_socket:send/2.